sync-engine refactor, separate base watcher functionality
